Break Theme Features into its own component
