+++ client/landing/stepper/declarative-flow/internals/steps-repository/design-setup/theme-features.scss
@@ -1,0 +1,19 @@
+.theme-features__features {
+    display: flex;
+    justify-content: flex-start;
+    flex-wrap: wrap;
+
+    .theme-features__feature {
+        display: inline-block;
+        height: 20px;
+        padding: 0 10px;
+        margin-right: 7px;
+        margin-top: 7px;
+        background: #dcdcde;
+        border-radius: 4px;
+        color: #2c3338;
+        font-weight: 500;
+        font-size: $font-body-small;
+        line-height: 20px;
+    }
+}
