+++ client/landing/stepper/declarative-flow/internals/steps-repository/design-setup/theme-features.tsx
@@ -1,0 +1,36 @@
+import classnames from 'classnames';
+import './theme-features.scss';
+
+interface Feature {
+	name: string;
+}
+
+interface FeaturesProps {
+	classNames?: string;
+	features: Feature[];
+	heading?: string;
+}
+
+const ThemeFeatures = ( { classNames, features, heading }: FeaturesProps ) => {
+	const classes = classnames( 'theme-features__features-wrap popup-item', classNames );
+	return (
+		<>
+			{ features.length > 0 && (
+				<div className={ classes }>
+					{ heading && `<h3>${ heading }</h3>` }
+					<div className="theme-features__features">
+						{ features.map( ( feature, idx ) => {
+							return (
+								<div className="theme-features__feature" key={ idx }>
+									{ feature.name }
+								</div>
+							);
+						} ) }
+					</div>
+				</div>
+			) }
+		</>
+	);
+};
+
+export default ThemeFeatures;
